% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{buildCdmSource}
\alias{buildCdmSource}
\title{buildCdmSource}
\usage{
buildCdmSource(sourceKey, sourceName = NULL, dbms = NULL,
  cdmDatabaseSchema = NULL, resultsDatabaseSchema = NULL,
  vocabDatabaseSchema = NULL, connectionString = NULL, sourceId = NULL,
  priority = 0, user = NULL, password = NULL, connectionDetails = NULL)
}
\arguments{
\item{sourceKey}{A unique string identifier for the data source}

\item{sourceName}{The name of the CDM source}

\item{dbms}{The dbms of the CDM source}

\item{cdmDatabaseSchema}{The name of the schema where CDM data sits}

\item{resultsDatabaseSchema}{The name of the schema where results data sits}

\item{vocabDatabaseSchema}{The name of the schema where the vocabulary sits}

\item{connectionString}{The JDBC connection string to the CDM source; REQUIRED if trying to insert a CDM source 
(\code{\link{insertCdmSources}}), unnecessary when trying to remove a CDM source 
(\code{\link{removeCdmSources}})}

\item{sourceId}{An optional value to pre-specify the database index identifier for the data source. 
If left NULL, this is either auto-generated by the Repo database, or the next available
source_id is obtained.}

\item{priority}{1 = the source become the one Atlas uses by default (e.g. to show record counts)
0 = the source will not become the one Atlas uses by default (e.g. to show record counts)}

\item{user}{(OPTIONAL) The user name for the connection to the CDM. Only used if encryption of source credentials is enabled.}

\item{password}{(OPTIONAL) The password for the connection to the CDM. Only used if encryption of source credentials is enabled.}

\item{connectionDetails}{A connectionDetails object created using DatabaseConnector of this source}
}
\value{
A CDM source object with the necessary parameters to fulfill SOURCE and SOURCE_DAIMON
}
\details{
Creates a CDM source object for use in the package with necessary connection information.
If using \code{\link{insertCdmSources}} then connectionString will need to be defined.
}
\examples{
## When building a CDM source to insert, and to make it top priority:

buildCdmSource(sourceKey = "TestDb_V123", sourceName = "Test Database v123", dbms = "redshift", cdmDatabaseSchema = "cdm",
resultsDatabaseSchema = "ohdsi_results", vocabDatabaseSchema = "cdm", priority = 1,
connectionString = "jdbc:redshift://some-redshift-cluster:5439/testdb_v123?user=some-user&password=some-password")
}
\author{
Ajit Londhe
}
