% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{createNetworkSource}
\alias{createNetworkSource}
\title{Creates network level CDM source for Atlas}
\usage{
createNetworkSource(cdmSources, networkConnectionDetails,
  networkDatabaseSchema)
}
\arguments{
\item{cdmSources}{The list of databases to cut over}

\item{networkConnectionDetails}{A connectionDetails object for the network source}

\item{networkDatabaseSchema}{The fully qualified name of the schema that will serve as the Network Source in Atlas}
}
\description{
Creates network level CDM source for Atlas
}
