% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{createOhdsiResultsTables}
\alias{createOhdsiResultsTables}
\title{createOhdsiResultsTables}
\usage{
createOhdsiResultsTables(cdmSources, baseUrl,
  initConceptHierarchy = FALSE, sqlOnly = FALSE)
}
\arguments{
\item{cdmSources}{The list of databases to cut over}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://api.ohdsi.org:80/WebAPI".}

\item{initConceptHierarchy}{Should the concept hierarchy be initializted?}

\item{sqlOnly}{Generate SQL only, don't execute}
}
\value{
none
}
\description{
createOhdsiResultsTables
}
\details{
Creates all the OHDSI Results tables needed by Atlas by pulling results DDL from WebAPI
}
\author{
Ajit Londhe
}
