% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{insertCdmSources}
\alias{insertCdmSources}
\title{insertCdmSources}
\usage{
insertCdmSources(repoConnectionDetails, cdmSources, sourceIdx = FALSE,
  daimonIdx = FALSE, sqlOnly = FALSE)
}
\arguments{
\item{repoConnectionDetails}{A ConnectionDetails object for the OHDSI Repository}

\item{cdmSources}{A list of CDM sources, built using CDM source object from \code{\link{buildCdmSource}}}

\item{sourceIdx}{Boolean to indicate if the SOURCE table uses an identity index for source_id}

\item{daimonIdx}{Boolean to indicate if the SOURCE_DAIMON table uses an identity index for source_daimon_id}

\item{sqlOnly}{Generate SQL only, don't execute}
}
\description{
insertCdmSources
}
\details{
Inserts CDM sources into OHDSI Repository. Creates new records in SOURCE and associated records in SOURCE_DAIMON.
As these tables may have identity indexes, boolean settings can be used to tell the function to skip instantiating new IDs.
}
\author{
Ajit Londhe
}
