% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{refreshAtlasSources}
\alias{refreshAtlasSources}
\title{Refresh CDM sources in Atlas}
\usage{
refreshAtlasSources(webApiUrl, webApiPort, useHttps = FALSE)
}
\arguments{
\item{webApiUrl}{The URL of the WebAPI server}

\item{webApiPort}{The port that WebAPI communicates on}

\item{useHttps}{Does the WebAPI endpoint utilize HTTPS?}
}
\description{
Refresh CDM sources in Atlas
}
\details{
Refreshes the source cache that feeds into Atlas. This step is necessary when making changes to the SOURCE and SOURCE_DAIMON tables,
so that the changes propagate to Atlas.
}
\author{
Ajit Londhe
}
