% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{removeCdmSources}
\alias{removeCdmSources}
\title{removeSources}
\usage{
removeCdmSources(repoConnectionDetails, cdmSources, sqlOnly = FALSE)
}
\arguments{
\item{repoConnectionDetails}{A ConnectionDetails object for the OHDSI Repository}

\item{cdmSources}{A list of CDM sources, built using CDM source object from \code{\link{buildCdmSource}}}

\item{sqlOnly}{Generate SQL only, don't execute}
}
\details{
Removes CDM sources and their cohort results from OHDSI Repository. Sources are removed
from the SOURCE and SOURCE_DAIMON table, and cohort results for those sources are cleared.
}
\author{
Ajit Londhe
}
