% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{renameSourceKey}
\alias{renameSourceKey}
\title{Rename the source key of an existing source}
\usage{
renameSourceKey(oldSourceKey, newSourceKey, repoConnectionDetails, sqlOnly)
}
\arguments{
\item{oldSourceKey}{The name of the old source key}

\item{newSourceKey}{The new name of the source key}

\item{repoConnectionDetails}{A ConnectionDetails object for the OHDSI Repository}

\item{sqlOnly}{Generate SQL only, don't execute}
}
\description{
Rename the source key of an existing source
}
