% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmAtlasCutover.R
\name{createOhdsiResultsTables}
\alias{createOhdsiResultsTables}
\title{createOhdsiResultsTables}
\usage{
createOhdsiResultsTables(cdmSources, sqlOnly = FALSE)
}
\arguments{
\item{cdmSources}{The list of databases to cut over}

\item{sqlOnly}{Generate SQL only, don't execute}
}
\value{
none
}
\details{
Creates all the OHDSI Results tables needed by Atlas
}
\author{
Ajit Londhe
}
